# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/api/countries": {
      "get": {
        "operationId": "CountriesController_getCountries",
        "summary": "Retrieve a paginated list of all the countries in the world",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of data per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "current page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "region",
            "required": false,
            "in": "query",
            "description": "Region to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Countries"
        ]
      }
    },
    "/v1/api/countries/{country}": {
      "get": {
        "operationId": "CountriesController_getCountry",
        "summary": "Retrieve detailed information for a specific country, including its languages,\npopulation, area, and bordering countries.",
        "parameters": [
          {
            "name": "country",
            "required": true,
            "in": "path",
            "description": "Country name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Countries"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shortlet API",
    "description": "OpenAPI swagger documentation for Shortlet backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}